significant acceleration of the downloading process of data (#24)
Thanks to a tip by @richardalligier

Adding the `-B` option to the ssh call deactivates the pretty-printing process and improves the processing and downloading time of data from the OpenSky Impala server.

A down side of this option is the disappearing of column names in the cache files. A new parameter has been added with the proper list of column names (variable is often named `process_columns` in the code) in order to help the `pd.read_csv` process.

In current examples, we have managed to properly infer the column names but it may be a bit dangerous in the future.